Adds stop_on_error config option
When set true, any HTTPError or HTTPConnectionError that is caught during the
vault read will be raised as a Puppet::DataBinding::LookupError and stop the
lookup instead of returning not_found.

Note that the vault gem returns nil in the case of a 404 therefore an error
isn't raised for a path not found even with this option true.